# Conflicts:
#	pychron/canvas/canvas2D/scene/yaml_scene_loader.py
#	pychron/database/core/database_adapter.py
#	pychron/pipeline/plot/overlays/spectrum.py
#	pychron/spectrometer/isotopx/spectrometer/ngx.py

+++ pychron/processing/analyses/view/magnitude_editor.py
@@ -68,7 +68,7 @@ class MagnitudeRenderer(TableDelegate):
         rect = option.rect
         w = rect.width()
 
-        rect.setWidth(w * min(1.0, v))
+        rect.setWidth(int(w * min(1.0, v)))
         painter.fillRect(rect, brush)
         painter.restore()
